[nasa/Crypotlib#147] UT passes - resolved issue where local variable was not appropriately initialized.

+++ include/crypto.h
@@ -209,10 +209,10 @@ int32_t Crypto_Check_Anti_Replay_Verify_Pointers(SecurityAssociation_t* sa_ptr, 
 int32_t Crypto_Check_Anti_Replay_ARSNW(SecurityAssociation_t* sa_ptr, uint8_t* arsn, int8_t* arsn_valid);
 int32_t Crypto_Check_Anti_Replay_GCM(SecurityAssociation_t* sa_ptr, uint8_t* iv, int8_t* iv_valid);
 
-// SDLS EP Related Functions
+// SDLS Related Functions
 extern uint8_t Crypto_Prep_Reply(uint8_t* ingest, uint8_t appID);
 extern void Crypto_Print_Sdls_Ep_Reply(void);
-extern uint16_t Crypto_Get_Sdls_Ep_Reply(uint8_t* buffer, uint16_t* length);
+extern int32_t Crypto_Get_Sdls_Ep_Reply(uint8_t* buffer, uint16_t* length);
 
 // Key Management Functions
 int32_t Crypto_Key_OTAR(void);
